pyproject.toml
[tool.poetry]
name = "spotter"
version = "0.1.0"
description = ""
authors = ["Mwapsam <mwapsam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^25.1"
django-types = "^0.20.0"
django-stubs-ext = "^5.1.3"
psycopg2 = "^2.9.10"
django-cors-headers = "^4.7.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
